+++ modules/alloy/src/test/ts/atomic/position/view/BounderCalcRepositionTest.ts
@@ -1,60 +1,32 @@
-import { UnitTest } from '@ephox/bedrock-client';
-import Jsc from '@ephox/wrap-jsverify';
+import { describe, it } from '@ephox/bedrock-client';
+import { assert } from 'chai';
+import * as fc from 'fast-check';
 
-import { bounds as makeBounds } from 'ephox/alloy/alien/Boxes';
+import * as Boxes from 'ephox/alloy/alien/Boxes';
 import * as Bounder from 'ephox/alloy/positioning/view/Bounder';
+import { assertInBounds, boundsArb, boxAndBoundsArb, boxArb } from 'ephox/alloy/test/BoundsUtils';
 
-UnitTest.test('BounderCalcRepositionTest', () => {
-
-  const maxBounds = 2000;
-  const minBounds = 0;
-  const zeroableArb = Jsc.integer(minBounds, maxBounds);
-
-  const arbTestCase = Jsc.bless({
-    generator: zeroableArb.generator.flatMap(
-      (boundsX: number) => zeroableArb.generator.flatMap(
-        (boundsY: number) => Jsc.integer(boundsX, maxBounds).generator.flatMap(
-          (newX: number) => Jsc.integer(boundsY, maxBounds).generator.flatMap(
-            (newY: number) => zeroableArb.generator.flatMap(
-              (width: number) => zeroableArb.generator.flatMap(
-                (height: number) => Jsc.integer(newX + width, maxBounds).generator.flatMap(
-                  (boundsW: number) => Jsc.integer(newY + height, maxBounds).generator.map(
-                    (boundsH: number) => ({
-                      newX,
-                      newY,
-                      width,
-                      height,
-                      boundsX,
-                      boundsY,
-                      boundsW,
-                      boundsH
-                    })
-                  )
-                )
-              )
-            )
-          )
-        )
-      )
-    )
+describe('BounderCalcRepositionTest', () => {
+  it('the repositioned box should always be within the bounds', () => {
+    fc.assert(fc.property(boxAndBoundsArb(-2000, 2000), ({ bounds, box }) => {
+      const output = Bounder.calcReposition(box, bounds);
+      assertInBounds(output, bounds);
+    }));
   });
 
-  Jsc.property(
-    'Check that all values have something visible within bounds',
-    arbTestCase,
-    (input: { newX: number; newY: number; width: number; height: number; boundsX: number; boundsY: number; boundsW: number; boundsH: number}) => {
-      const bounds = makeBounds(input.boundsX, input.boundsY, input.boundsW, input.boundsH);
-      const output = Bounder.calcReposition(input.newX, input.newY, input.width, input.height, bounds);
+  it('a box inside the bounds should be the same size', () => {
+    const bounds = Boxes.bounds(-200, -200, 400, 400);
+    fc.assert(fc.property(boxArb(-199, 199), (box) => {
+      const output = Bounder.calcReposition(box, bounds);
+      assert.deepEqual(output, box);
+    }));
+  });
 
-      const xIsVisible = output.limitX <= (bounds.right - input.width) && output.limitX >= bounds.x;
-      const yIsVisible = output.limitY <= (bounds.bottom - input.height) && output.limitY >= bounds.y;
-      if (!xIsVisible) {
-        return 'X is not inside bounds. Returned: ' + JSON.stringify(output);
-      } else if (!yIsVisible) {
-        return 'Y is not inside bounds. Returned: ' + JSON.stringify(output);
-      } else {
-        return true;
-      }
-    }
-  );
+  it('a box outside the bounds should be shrunk to the size of the bounds', () => {
+    const box = Boxes.bounds(0, 0, 400, 400);
+    fc.assert(fc.property(boundsArb(50, 350), (bounds) => {
+      const output = Bounder.calcReposition(box, bounds);
+      assert.deepEqual(output, bounds);
+    }));
+  });
 });